refactor: made a few changes based on feedback
